+++ CDS-APP-2-iOS/Presentation/Category/Model/Category.swift
@@ -1,0 +1,13 @@
+//
+//  Category.swift
+//  CDS-APP-2-iOS
+//
+//  Created by 변희주 on 2023/11/20.
+//
+
+import UIKit
+
+struct Category {
+    let image: UIImage
+    let label: String
+}
