+++ CDS-APP-2-iOS/Presentation/Category/Model/CategoryDetailList.swift
@@ -1,0 +1,12 @@
+//
+//  CategoryDetailList.swift
+//  CDS-APP-2-iOS
+//
+//  Created by 변희주 on 2023/11/21.
+//
+
+import UIKit
+
+struct CategoryDetailList {
+    let label: String
+}
