+++ CDS-APP-2-iOS/Presentation/Category/Model/CategoryList.swift
@@ -1,0 +1,12 @@
+//
+//  CategoryList.swift
+//  CDS-APP-2-iOS
+//
+//  Created by 변희주 on 2023/11/21.
+//
+
+import UIKit
+
+struct CategoryList {
+    let label: String
+}
